+++ addon/components/jubilee-modal.ts
@@ -1,14 +1,18 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 
-export default class JubileeModalComponent extends Component {
+interface JubileeModalArgs {
+  maxContent?: boolean;
+}
+
+export default class JubileeModal extends Component<JubileeModalArgs> {
   /**
    * maxContent: bool(default true)
    */
 
   // #region Properties
 
-  get isMax() {
+  get isMax(): boolean {
     return (
       typeof this.args.maxContent === 'undefined' ||
       this.args.maxContent === true