+++ CHANGELOG.md
@@ -6,7 +6,7 @@ An online documentation of the components is available [here](https://ui.jubilee
 
 ##### Added
 
-- [#12](https://github.com/jubileesoft/ember-jubileesoft-ui/pull/11) Add light and dark theme to JubileeInput.
+- [#11](https://github.com/jubileesoft/ember-jubileesoft-ui/pull/11) Add light and dark theme to JubileeInput.
 
 ##### Changed
 
@@ -19,6 +19,10 @@ An online documentation of the components is available [here](https://ui.jubilee
   - Bump uuidv4 from 6.0.0 to 6.0.7
   - Bump sass from 1.22.10 to 1.26.5
 - [#8](https://github.com/jubileesoft/ember-jubileesoft-ui/pull/8) Upgrade to ember-cli@3.17.0
+
+##### Fixed
+
+ - [#12](https://github.com/jubileesoft/ember-jubileesoft-ui/pull/12) Fix docs for JubileeInput and add missing attribute @placeholder.
 
 ## <span style="color: #0366d6;">1.0.1</span>